requirements.txt
av==10.0.0
fastapi==0.95.2
numpy==1.24.3
opencv_python==4.6.0.66
pandas==1.5.1
plotly==5.10.0
pydantic==1.9.1
requests==2.28.1
streamlit==1.22.0
streamlit_server_state==0.16.0
streamlit_wallet_connect==0.1.0
streamlit_webrtc==0.45.0
typing_extensions==4.5.0
uvicorn==0.22.0
web3==6.0.0b6
